all expected use case and test cased completed.

+++ src/main/java/com/example/deliveryparsing/report/dtos/AggregatedDeliveryItem.java
@@ -7,9 +7,9 @@ import lombok.Data;
 @Data
 public class AggregatedDeliveryItem {
   private Placement placement;
-  private BigInteger deliveryTotalImpressions;
+  private long deliveryTotalImpressions;
 
-  public AggregatedDeliveryItem(Placement placement, BigInteger deliveryTotalImpressions) {
+  public AggregatedDeliveryItem(Placement placement, long deliveryTotalImpressions) {
     this.placement = placement;
     this.deliveryTotalImpressions = deliveryTotalImpressions;
   }
